backend/requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.30.6
# python-multipart==0.0.9
# vosk==0.3.45
# requests==2.32.3
# gTTS==2.5.3
# ffmpeg-python==0.2.0
# pydub==0.25.1
# numpy==2.1.1
# python-dotenv==1.0.1



fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
vosk==0.3.45
requests==2.32.3
gTTS==2.5.3
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==2.1.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
argostranslate==1.9.1
